Rename urls.js to scopes.js
