+++ packages/php-wasm-browser/src/emscripten-download-monitor.js
@@ -1,4 +1,4 @@
-import { DEFAULT_BASE_URL } from "./urls";
+import { DEFAULT_BASE_URL } from "./";
 
 /*
  * An approximate total file size to use when the actual
